Fix: not accessing db collection when db is not ready (#1421)

+++ src/Component/backupper/mongodb.ts
@@ -115,10 +115,14 @@ export class MongoBackupper extends Backupper {
   }
 
   protected async deleteGuildData(guildId: string){
-    Promise.allSettled([
-      this.collections.queue.deleteOne({ guildId }),
-      this.collections.status.deleteOne({ guildId }),
-    ]).catch(this.logger.error);
+    if(this.collections && this.dbConnectionReady){
+      Promise.allSettled([
+        this.collections.queue.deleteOne({ guildId }),
+        this.collections.status.deleteOne({ guildId }),
+      ]).catch(this.logger.error);
+    }else{
+      this.logger.warn(`No data was removed (guildId: ${guildId}) due to no connection`);
+    }
   }
 
   @timeLoggedMethod
